Renamed areagraphreact -> areagraph
