+++ web/src/scss/content/panel/_country-panel.scss
@@ -143,7 +143,7 @@
     }
 }
     
-#country-history-mix-react {
+#country-history-mix {
     height: 10rem;
 
     .layer path {
